Move modules and cost_aggregator.py out of api submodule
