+++ app/api/another_module.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, Field
 
-from app.api.module_interface import ModuleInterface
+from app.module_interface import ModuleInterface
 
 
 class AnotherModuleRequest(BaseModel):
